backend/services/retrieval/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
pydantic
python-dotenv
psycopg2-binary
pgvector
numpy
